simplified setup

+++ setup.ps1
@@ -1,4 +1,14 @@
-uv pip install -r requirements.txt
-uv pip install --pre torch torchvision --index-url "https://download.pytorch.org/whl/nightly/cu128"
-C:\Users\g2-leonovs\repo\remote-main\rclone\rclone.exe copy gdrive:datasets factual_models/
-explorer.exe factual_models;
+$data = "tinystories"
+
+if (-not (Test-Path -Path $data)) {
+	uv pip install -r requirements.txt
+	uv pip install --pre torch torchvision --index-url "https://download.pytorch.org/whl/nightly/cu128"
+	echo "[rclone] downloading data"
+	New-Item -ItemType Directory -Path $data -Force
+	C:\Users\g2-leonovs\repo\remote-main\rclone\rclone.exe copy gdrive:datasets/$data factual_models/$data
+}
+echo "[python] initializing model"
+# explorer.exe factual_models;
+python start.py
+C:\Users\g2-leonovs\repo\remote-main\rclone\rclone.exe copy results gdrive:results
+timeout /t 20; shutdown /l
